In from_geofile(), only delete the .xml.gz file if it didn't exist previously (otherwise this might result in unpleasant surprises for the user).

+++ src/finmag/util/meshes.py
@@ -55,13 +55,16 @@ def from_geofile(geofile, save_result=True):
 
     """
     result_filename = os.path.splitext(geofile)[0] + ".xml.gz"
+    result_file_already_exists = False
     if os.path.isfile(result_filename) and os.path.getctime(result_filename) > os.path.getctime(geofile):
+        result_file_already_exists = True
         logger.debug("The mesh %s already exists, and is automatically returned." % result_filename)
     else:
         result_filename = compress(convert_diffpack_to_xml(run_netgen(geofile)))
 
     mesh = Mesh(result_filename)
-    if not save_result:
+    if not save_result and not result_file_already_exists:
+        # We only delete the .xml.gz file if it didn't exist previously
         os.remove(result_filename)
     return mesh
